### 2018-01-24 #188 @marcosci
NLMR
-   Provides neutral landscape models and utility functions to work with and extend them. NLMR uses raster as its geospatial framework and can therefore easily be used in landscape analysis to test against the influence of landscape patterns and serve as a basis for spatial simulation modeling.

```
Package: NLMR
Title: Simulating Neutral Landscape Models
Version: 0.2
Authors@R: c(person("Marco", "Sciaini", email = "sciaini.marco@gmail.com", role = c("aut", "cre")),
    person("Matthias", "Fritsch", email = "matthias.fritsch@forst.uni-goettingen.de", role =  c("aut")),
    person("Craig", "Simpkins", email = "simpkinscraig063@gmail.com", role =  c("aut")),
    person("Cédric", "Scherer", email = "cedricphilippscherer@gmail.com", role =  c("ctb"), comment = "Implemented nlm_neigh"))
Description: Provides neutral landscape models
    (Gardner et al. 1987 <doi:10.1007/BF02275262>,
    With 1997 <doi:10.1046/j.1523-1739.1997.96210.x>) that can easily extend in
    existing landscape analyses. Neutral landscape models range from "hard"
    neutral models (completely random distibuted), to "soft" neutral models (definable spatial characteristics) and
    generate landscape patterns that are independent of ecological processes.
    Thus, these patterns can be used as null models in landscape ecology. 'NLMR'
    combines a large number of algorithms from other published software for simulating neutral landscapes (Saura &
    Martínez 2000   <doi:10.1023/A:1008107902848>, Etherington et
    al. 2015 <doi:10.1111/2041-210X.12308>) and
    includes utility functions to classify and combine the multiple landscapes. The
    simulation results are obtained in a geospatial data format (raster* objects
    from the 'raster' package) and can, therefore, be used
    in any sort of raster data operation that is performed with standard
    observation data.
Copyright: file inst/COPYRIGHTS
    checkmate,
    dismo,
    igraph,
    RandomFields,
    raster,
    rasterVis,
    sp,
    spatstat,
    viridis,
    extrafont
    covr,
    lintr
    knitr
```

    - [https://github.com/marcosci/NLMR](https://github.com/marcosci/NLMR)

    - ***geospatial data***, because the package simulates land use patterns in a geospatial format.
    - ***Exploratory data analysis***, because the package can be used as a very fundamental null hypothesis in any analysis that formulate questions around the influence of landscape pattern on processes (e.g. ecological, social processes),

    - ***Target audience***: Mainly ecologists, since neutral landscape models are a well-established method in their field. But the package itself can be used by a broad audience, as it is convenient way to quickly prototype spatial simulation models.
    - ***Scientific applications***: Formulating null hypotheses in a landscape analysis, test and develop landscape metrics, gradient analysis of the influence of landscape patterns on ecological processes, starting point for spatial simulations, such as agent-based models.

    - There was ecomodtools, which could simulate a small fraction of the neutral landscape models in NLMR. But it is not under active development anymore and does not compile under newer R versions.

- [x] Do you intend for this package to go on CRAN?  (it already is)
	- [x] You intend to archive the code for the package in a long-term repository which meets the requirements of the journal.

    - No errors/warnings
