### 2018-01-24 #188 @marcosci
Hi Folks,

was quite unsure about the fit of our package into the rOpenSci world - but we would love to get your review and platform.

Best wishes,
Marco
