### 2018-01-29 marcosci/NLMR#10 @marcosci
Metrics
Implement more metric functions

- Fractal dimension
- Hurst
- Lacunarity
- ... ?
