### 2018-01-30 95849f3
implement util_facetplot
code a prototype skeleton for facet plotting a rasterlayer, rasterstack, rasterbrick or any list with raster maps side by side.

mmmh, compare this beauty against rasterVis::levelplot()!
