### 2018-01-31 7f6363b
fix lintr-bot comments
fix auto indented lines, inserted whitespace and quotes.
