### 2018-01-31 #188 @maelle
### Editor checks:

- [x ] **Fit**: The package meets criteria for [fit](policies.md#fit) and [overlap](policies.md#fit)
- [ x] **Automated tests:** Package has a testing suite and is tested via Travis-CI or another CI service.
- [ x] **License:** The package has a CRAN or OSI accepted license
- [ x] **Repository:** The repository link resolves correctly

---

#### Editor comments

Thanks for your submission @marcosci! I'm currently looking for reviewers. I have these comments/questions:

* `goodpractice::gp()` returned "♥ Ahh! Rad package! Keep up the tiptop work!" which I had to pass on. :wink:

* the copyright file is for a font, where is this font used?

* please add this badge to the repo README

---

Reviewers:
Due date:
