### 2018-01-31 #188 @marcosci
Hey @maelle and @sckott,

thanks a lot for the effort :)

The font is used in the theme function (`theme_nlm`) and can be imported via
`util_import_roboto_condensed`.

Which badge? :) There seems to be something missing in your reply.
