### 2018-01-31 ae1eaf3
include ropensci badge
