+++ README.md
@@ -30,8 +30,6 @@ Here we will provide a simple example on using `NLMR`:
 
 ``` r
 library(NLMR)
-#> Registering PDF &
-#>                                            PostScript fonts with R
 library(magrittr)
 library(ggplot2)  # to extend the plot functionality of NLMR 
 library(SDMTools) # to calculate basic landscape metrics
@@ -69,9 +67,9 @@ raster::as.matrix(nlm_raster) %>%
 
 |  patchID|  n.cell|  n.core.cell|  n.edges.perimeter|  n.edges.internal|  area|  core.area|  perimeter|  perim.area.ratio|  shape.index|  frac.dim.index|  core.area.index|
 |--------:|-------:|------------:|------------------:|-----------------:|-----:|----------:|----------:|-----------------:|------------:|---------------:|----------------:|
-|        0|    1249|          519|                826|              4170|  1249|        519|        826|         0.6613291|     5.816901|        1.495155|        0.4155324|
-|        1|     615|          192|                570|              1890|   615|        192|        570|         0.9268293|     5.700000|        1.544576|        0.3121951|
-|        2|     636|          215|                542|              2002|   636|        215|        542|         0.8522013|     5.313726|        1.520936|        0.3380503|
+|        0|    1248|          560|                784|              4208|  1248|        560|        784|         0.6282051|     5.521127|        1.480683|        0.4487179|
+|        1|     581|          149|                630|              1694|   581|        149|        630|         1.0843373|     6.428571|        1.589827|        0.2564544|
+|        2|     671|          226|                586|              2098|   671|        226|        586|         0.8733234|     5.634615|        1.532402|        0.3368107|
 
 Citation
 --------
