### 2018-02-01 89925cf
add doc and tests
add documentation text and testfile for automated tests while building package.
