### 2018-02-02 bc78b6f
update util_facetplot
