+++ NAMESPACE
@@ -24,6 +24,7 @@ export(theme_nlm_grey_discrete)
 export(util_binarize)
 export(util_calc_boundaries)
 export(util_classify)
+export(util_facetplot)
 export(util_import_roboto_condensed)
 export(util_merge)
 export(util_plot)
