+++ DESCRIPTION
@@ -44,6 +44,7 @@ Imports:
     spatstat,
     stats,
     tibble,
+    tidyr,
     viridis,
     extrafont
 URL: https://marcosci.github.io/NLMR/
@@ -56,34 +57,3 @@ Suggests:
     lintr
 VignetteBuilder:
     knitr
-Collate: 
-    'metric_area.R'
-    'nlm_curds.R'
-    'nlm_distancegradient.R'
-    'nlm_edgegradient.R'
-    'util_rescale.R'
-    'nlm_fBm.R'
-    'nlm_gaussianfield.R'
-    'nlm_mosaicfield.R'
-    'nlm_mpd.R'
-    'nlm_neigh.R'
-    'nlm_percolation.R'
-    'nlm_planargradient.R'
-    'nlm_polylands.R'
-    'nlm_random.R'
-    'nlm_randomcluster.R'
-    'nlm_randomelement.R'
-    'nlm_randomrectangularcluster.R'
-    'nlm_wheys.R'
-    'package.R'
-    'theme_nlm.R'
-    'util_binarize.R'
-    'util_calcboundaries.R'
-    'util_classify.R'
-    'util_import_roboto_condensed.R'
-    'util_merge.R'
-    'util_plot.R'
-    'util_raster2tibble.R'
-    'util_tibble2raster.R'
-    'util_w2cp.R'
-    'zzz.R'
