### 2018-02-03 marcosci/NLMR#10 @csim063
Useful paper to get thinking about which metrics may be most useful. https://www.sciencedirect.com/science/article/pii/S1470160X07001306
