### 2018-02-03 marcosci/NLMR#10 @marcosci
* Proportion of landscape
* Number of instances
* Instance size distribution
* Instance perimeter
* Instance bounding box aspect-ratio and proportion of bounding box area
* Instance edge contrast index
* Total edge

> Slager CTJ, De Vries B. 2013. Landscape generator: Method to generate landscape configurations for spatial plan-making. Computers, Environment and Urban Systems 39:1–11. Elsevier Ltd.
