### 2018-02-13 b0a00ed
update Rbuildignore
