### 2018-02-14 e4a7699
update readme and pkgdown
