### 2018-02-16 0911e40
fix errors/warnings from rhub/win_builder
