+++ R/zzz.R
@@ -8,14 +8,17 @@
   # https://github.com/hrbrmstr/hrbrthemes
 
   if (.Platform$OS.type == "windows")  {
-    if (interactive()) packageStartupMessage("Registering Windows fonts with R")
+    if (interactive()) packageStartupMessage(
+      "Registering Windows fonts with R for advanced plotting options"
+      )
     # work around for https://github.com/wch/extrafont/issues/44
     windowsFonts <- grDevices::windowsFonts
     extrafont::loadfonts("win", quiet = TRUE)
   }
 
-  if (interactive()) packageStartupMessage("Registering PDF &
-                                           PostScript fonts with R")
+  if (interactive()) packageStartupMessage(
+      "Registering PDF & PostScript fonts with R for advanced plotting options"
+      )
   #work around for https://github.com/wch/extrafont/issues/44
   pdfFonts <- grDevices::pdfFonts
   extrafont::loadfonts("pdf", quiet = TRUE)
