+++ man/theme_nlm.Rd
@@ -8,54 +8,46 @@
 \alias{theme_nlm_grey_discrete}
 \title{theme_nlm}
 \usage{
-theme_nlm(base_family = "Roboto Condensed", base_size = 11.5,
-  plot_title_family = base_family, plot_title_size = 18,
-  plot_title_face = "bold", plot_title_margin = 10, subtitle_family = if
-  (.Platform$OS.type == "windows") "Roboto Condensed" else
-  "Roboto Condensed Light", subtitle_size = 13, subtitle_face = "plain",
+theme_nlm(base_family = "serif", base_size = 11.5,
+  plot_title_family = "serif", plot_title_size = 18,
+  plot_title_face = "bold", plot_title_margin = 10,
+  subtitle_family = "serif", subtitle_size = 13, subtitle_face = "plain",
   subtitle_margin = 15, strip_text_family = base_family,
-  strip_text_size = 12, strip_text_face = "plain", caption_family = if
-  (.Platform$OS.type == "windows") "Roboto Condensed" else
-  "Roboto Condensed Light", caption_size = 9, caption_face = "plain",
+  strip_text_size = 12, strip_text_face = "plain",
+  caption_family = "serif", caption_size = 9, caption_face = "plain",
   caption_margin = 10, legend_title = "Z",
   plot_margin = ggplot2::margin(30, 30, 30, 30), ratio = 1,
   viridis_scale = "D")
 
-theme_nlm_discrete(base_family = "Roboto Condensed", base_size = 11.5,
+theme_nlm_discrete(base_family = "serif", base_size = 11.5,
   plot_title_family = base_family, plot_title_size = 18,
-  plot_title_face = "bold", plot_title_margin = 10, subtitle_family = if
-  (.Platform$OS.type == "windows") "Roboto Condensed" else
-  "Roboto Condensed Light", subtitle_size = 13, subtitle_face = "plain",
+  plot_title_face = "bold", plot_title_margin = 10,
+  subtitle_family = "serif", subtitle_size = 13, subtitle_face = "plain",
   subtitle_margin = 15, strip_text_family = base_family,
-  strip_text_size = 12, strip_text_face = "plain", caption_family = if
-  (.Platform$OS.type == "windows") "Roboto Condensed" else
-  "Roboto Condensed Light", caption_size = 9, caption_face = "plain",
+  strip_text_size = 12, strip_text_face = "plain",
+  caption_family = "serif", caption_size = 9, caption_face = "plain",
   caption_margin = 10, legend_title = "Z", legend_labels = NULL,
   plot_margin = ggplot2::margin(30, 30, 30, 30), ratio = 1,
   viridis_scale = "D")
 
-theme_nlm_grey(base_family = "Roboto Condensed", base_size = 11.5,
+theme_nlm_grey(base_family = "serif", base_size = 11.5,
   plot_title_family = base_family, plot_title_size = 18,
-  plot_title_face = "bold", plot_title_margin = 10, subtitle_family = if
-  (.Platform$OS.type == "windows") "Roboto Condensed" else
-  "Roboto Condensed Light", subtitle_size = 13, subtitle_face = "plain",
+  plot_title_face = "bold", plot_title_margin = 10,
+  subtitle_family = "serif", subtitle_size = 13, subtitle_face = "plain",
   subtitle_margin = 15, strip_text_family = base_family,
-  strip_text_size = 12, strip_text_face = "plain", caption_family = if
-  (.Platform$OS.type == "windows") "Roboto Condensed" else
-  "Roboto Condensed Light", caption_size = 9, caption_face = "plain",
+  strip_text_size = 12, strip_text_face = "plain",
+  caption_family = "serif", caption_size = 9, caption_face = "plain",
   caption_margin = 10, legend_title = "Z",
[…]
