### 2018-02-17 #188 @laurajanegraham
Hi,

Thanks for the reminder.

I have put time aside Monday and Tuesday to get this review done.

Cheers,

Laura

On 16 Feb 2018 16:51, "Jeffrey W Hollister" <notifications@github.com>
wrote:

> Perfect and thank you.
>
> On Feb 16, 2018 11:35 AM, "Maëlle Salmon" <notifications@github.com>
> wrote:
>
> > @jhollist <https://github.com/jhollist> ok, thanks for the update! March
> > the 1st?
> >
> > —
> > You are receiving this because you were mentioned.
> > Reply to this email directly, view it on GitHub
> > <https://github.com/ropensci/onboarding/issues/188#
> issuecomment-366287026>,
> > or mute the thread
> > <https://github.com/notifications/unsubscribe-auth/
> AFL8SyqLmwOuTY6ufoDmLMZA4r31u6a4ks5tVa46gaJpZM4RrS2f>
> > .
> >
>
