### 2018-02-19 #188 @marcosci
Yup :) But develop should also be stable, master contains only the CRAN versions :)
