### 2018-02-19 #188 @marcosci
master makes more sense maybe, yes.
We plan to integrate the [ecomodtools ](https://rdrr.io/rforge/ecomodtools/)package and I am in contact with the author - but there is no schedule for that yet :)
