### 2018-02-23 #188 @laurajanegraham
Hi @maelle and @marcosci, here is my review. I really like this package, and reviewing it has given me a good opportunity to get to know the features in more depth, so thanks both :)

- [x] **A statement ofneed** clearly stating problems the software is designed to solve and its target audience in README

---

The package authors have provided a much needed collection of neutral landscape models for the R platform. I was pleased to see that it was not just a repackaging of the Python nlmpy package, but also provides additional neutral landscape models as well as some utility functions for visualising these landscapes. In terms of statement of need; the README provides a good explanation of the need for the package, but I would also add that it allows users to create NLMs within a self-contained, reproducible framework.

The code is mostly compact, efficient and follows the standards for rOpenSci packages. I have some concerns around the unit tests: the tests check that the output is of the correct form (class, size etc.), but do not go for edge cases, or try to break the functions. I have noted below some specific cases where functions either broke or did not work as expected.

One way in which I would suggest for improvement of the package would be to improve the documentation. It is mostly good, but the amount of information on the functions is variable. I have commented where I think specific functions could do with more documentation below. I do, however, see this as more of a long term thing, rather than something I'd expect to be improved as part of this review process.

The package name is sensible. It is, however, not in lower case as recommended by the packaging guide. The functions and variables all follow rOpenSci guidelines.

Below I have outlined some issues I found with some of the functions. I think most of these are pretty easy fixes with the exception of `nlm_fBm`: I'm not entirely sure that this is fixable, unless I've misunderstood one of the arguments to the underlying RandomFields function, and without being able to include the full range of values, I think it may be better to remove this function.

`metric_area`: This function will only give the values of either one class, or all classes. For example:

```r
library(NLMR)
x <- nlm_random(100, 100)
y <- c(0.5, 0.25, 0.25)
z <- util_classify(x, y)

metric_area(z, c(1,2))
```

```
## Error in `$<-.data.frame`(`*tmp*`, "class", value = c(1, 2)): replacement has 2 rows, data has 3
```

The problem is in the if(length(poi) == 1) statement. This is not necessary. Any input could be evaluated by the code inside of the if part. Additionally this outputs a dataframe for Proportion_Area, and not a tibble, as stated in the documentation.

----

`nlm_curds` & `nlm_wheys`:

- No `checkmate` tests at the beginning.

- There is borrowed code between these two. It might make sense to combine them (with a wheys T/F option), or to use the `nlm_curds` function inside `nlm_wheys` instead of repeating the code.

- These seem quite different in the way they are written to the other functions. Primarily in that the others all specify `nrow`, `ncol`, `resolution`, whereas these specify extent. Is it possible to harmonise these?

----

`nlm_edgegradient`: The call to `nlm_planargradient` (L66) is misspecified. It should be
`gradient_raster <- nlm_planargradient(ncol, nrow, direction = direction)`

or

`gradient_raster <- nlm_planargradient(ncol, nrow, resolution, direction)`

or change the order of the arguments to `nlm_planargradient`

----

`nlm_fbm`:

- Add in checkmate::assert_true(fract_dim > 0).
- The code checks that fract_dim is < 2, but the help file states that this parameter is including 2 - fract_dim = numeric in (0, 2]
- Values of `fract_dim` between ~1.56 and ~1.99 generate an error. This is caused by the `RandomFields::RFsimulate` function.

```r
fBm_raster  <- nlm_fBm(ncol = 20, nrow = 30, fract_dim = 1.9)
```

```
## Error in rfInit(model = list("Simulate", setseed = eval(parse(text = "quote(set.seed(seed=seed))")), : searching a simulation method for 'RMfbm':  Running out of list of methods. Are the RFoptions() too restrictive?
##  You get (more) internal information if you set RFoptions(cPrintlevel=3) before running your code.
```

- In the details for this function, you state *Implementation of this method is limited to landscapes with extents less than 90 by 90 cells.*, I would recommend also putting in an `assert_true` statement to this effect.

- If this function is run with a random seed, it holds onto it until a new seed is set. This can be fixed by removing the `if (!is.null(user_seed))` statement. The default `user_seed = NULL` will set the seed back to NA.

```r
nlm_fBm(ncol = 20, nrow = 30, fract_dim = 1, user_seed = 30)
```

```
## NOTE: simulation is performed with fixed random seed 30.
## Set 'RFoptions(seed=NA)' to make the seed arbitrary.
```

```
## class       : RasterLayer
## dimensions  : 30, 20, 600  (nrow, ncol, ncell)
## resolution  : 1, 1  (x, y)
## extent      : 0, 20, 0, 30  (xmin, xmax, ymin, ymax)
## coord. ref. : NA
## data source : in memory
## names       : layer
## values      : 0, 1  (min, max)
```

```r
nlm_fBm(ncol = 20, nrow = 30, fract_dim = 1)
```

```
## NOTE: simulation is performed with fixed random seed 30.
## Set 'RFoptions(seed=NA)' to make the seed arbitrary.
```

```
## class       : RasterLayer
## dimensions  : 30, 20, 600  (nrow, ncol, ncell)
## resolution  : 1, 1  (x, y)
## extent      : 0, 20, 0, 30  (xmin, xmax, ymin, ymax)
## coord. ref. : NA
## data source : in memory
## names       : layer
## values      : 0, 1  (min, max)
```

----

`nlm_gaussianfield`:

- See above issues with setting the random seed in `nlm_fBm`

- Setting the mean value of the field doesn't work because of the line `pred_raster <- pred_raster - raster::cellStats(pred_raster, "min")`. Is there a purpose for this line of code? Removing this produces Gaussian random field surfaces with the given mean.

- Add `lower = 0` to the checks for `resolution`, `mag_var` and `nug`

- Some additions to the documentation could improve the usability of this function. I would recommend further explanation of what each of the variance parameters do, in a way that the implications for the generated surface can be interpreted. My understanding of these are that:
  - `mag_var` sets the variation of the broad landscape
  - `nug` sets the variation within the scale of `autocorr_range`: low `nug` values will produce a highly spatially autocorrelated surface and high `nug` values a rougher surface.

- Missing a reference, which would be useful for users to get a better understanding of how this function works.

----

`nlm_mosaicfield`:

- The collect stage overwrites the first step, need to change the for loop on line 91 to `for (i in 2:n)` (can also remove the `i <- 2` from line 86)

- In addition, so more information about how these work, and when you would use them would be useful.

----

`nlm_mpd`:

- Generally really clearly explained - more detail on what effect changing rand_dev has on the end landscape would be useful. From some experimenting, it looks like it just changes the overall variance of the landscape.

----

`nlm_neigh`:

- Additional checks: set `lower` and `upper` values in the `assert_numeric` statement for `p_neigh` and `p_empty`; check neighborhood contains the accepted values; check proportions sums to 1.

- Note on the last one - it's important because if proportions do not sum to 1, then the remaining cells are assigned to class 0, meaning that `proportions = c(0.1, 0.1, 0.1)` would result in a landscape with 80% class 0 and 10% the other two.

- Because the categories are integers, it doesn't make sense to rescale the raster for this function. I would suggest defaulting to `FALSE` or getting rid of the option altogether.

- `rev(proportions)` on line 91 means that the proportions are assigned to the categories backwards. e.g. `nlm_neigh(ncol = 50, nrow = 50, p_neigh = 0.5, p_empty = 0.5, categories = 2, proportions = c(0.75, 0.25))` will result in a landscape where 25% is assigned to 0 and 75% to 1. This could be confusing when the proportions are closer.

----

`nlm_polylands`:

- I wonder if this would be simpler split into two functions for each of the options. I understand the thinking behind keeping them as one, but as there are no shared outputs, and few shared parameters, they may make more sense separately.

- Some more information about what each of the parameters does here would be useful. For example, parameter R for option 2 is described as the interaction radius. I assumed this was related to the size of the landscape, so put in a value of 3 (so assuming that cells within a distance of 3 cells from each other are interacting). This causes the function to hang. It's the `spatstat::rStrauss(200, gamma = g, R = R)` that is doing this, and I found that it wouldn't run for me for a value of R > 0.1 (using g = 0.5).

- like with `nlm_neigh`, rescale doesn't make sense here.

----

`nlm_randomcluster`:

- The neighbourhood input is an integer here, whereas it was method names in `nlm_neigh`. It would be helpful to unify this (and the spelling of neighbourhood - one US, one British English). My preference would be the numbers, because I always forget the names and it's less to type!

----

`nlm_randomrectangularcluster`:

- Should there also be a test that minl < maxl?

----

`util_calc_boundaries` and `util_w2cp`:
  should these be exported, or are they meant to be internal functions?

----

`util_facetplot`: is it possible to add a `nrow` or `ncol` option into here to control how the facet is laid out?

----
