### 2018-02-23 #188 @maelle
@laurajanegraham thanks a lot for your excellent review! 👌 👏

@marcosci you can keep the name of the package now that it's on CRAN, not a strict rule on our part. :-) you can wait for the second review before responding if you prefer!
