### 2018-02-23 #188 @marcosci
thanks @laurajanegraham for your effort - really appreciate it 👍
didn't find anything that should not be fixable in my head - will wait for Jeff and then go through your code and test comments.

and your definitely right - the documentation needs improvement ... we're on it. 🚶
