### 2018-02-23 #188 @jhollist
Top of my to-do list for next week and have already started it...

Like what I see so-far!

On Fri, Feb 23, 2018 at 11:19 AM, Marco Sciaini <notifications@github.com>
wrote:

> thanks @laurajanegraham <https://github.com/laurajanegraham> for your
> effort - really appreciate it 👍
> didn't find anything that should not be fixable in my head - will wait for
> Jeff and then go through your code and test comments.
>
> and your definitely right - the documentation needs improvement ... we're
> on it. 🚶
>
