### 2018-02-25 7cc2d50
update README and typos
