+++ R/nlm_edgegradient.R
@@ -18,7 +18,7 @@
 #'
 #' @details
 #' Simulates a linear gradient orientated on a specified or random direction
-#' that has a central peak. which runs perpendicular to the gradient direction.
+#' that has a central peak, which runs perpendicular to the gradient direction.
 #'
 #' @examples
 #'
