+++ R/nlm_mosaicfield.R
@@ -1,6 +1,6 @@
 #' nlm_mosaicfield
 #'
-#' @description Simulate a mosaic random field neutral landscape model.
+#' @description Simulates a mosaic random field neutral landscape model.
 #'
 #' @param ncol [\code{numerical(1)}]\cr
 #' Number of columns for the raster.
