+++ README.Rmd
@@ -29,7 +29,9 @@ knitr::opts_chunk$set(
 `NLMR` builds on the advantages of the ``raster``-package and returns all simulation as ``RasterLayer``-objects, thus ensuring a direct compability to common GIS tasks and a pretty flexible and simple usage.
 Furthermore, it simulates NLMs within a self-contained, reproducible framework.
 
- <img src="vignettes/nlmr_grid.png" align="right"  width="100%" />
+<img src="vignettes/nlmr_grid.png"  width="100%" />
+
+<br>
 
 ## Installation
 
