### 2018-02-28 #188 @jhollist
Here is my review!  In short, I like the package and hope to find some uses for it.  I also have review envy.  Really nice job on your review, @laurajanegraham !!!

Any questions, let me know.

>##### For packages co-submitting to JOSS
>
>- [ ] The package has an **obvious research application** according to [JOSS's definition](http://joss.theoj.org/about#submission_requirements)
>
>The package contains a `paper.md` matching [JOSS's requirements](http://joss.theoj.org/about#paper_structure) with:
>
>- [ ] **A short summary** describing the high-level functionality of the software
>- [ ] **Authors:**  A list of authors with their affiliations
>- [ ] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [ ] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

~ 8 hours

The `NLMR` package provides a suite of functions to simulate neutral landscapes with multiple algorithms plus a set of utility functions for working with the resulting `raster` objects.  As a landscape ecologist who cut his teeth when many of these neutral landscapes were being developed, seeing them implemented as an R package was exciting.  Their implementation in `NLMR` is nice and the functions are straightforward to use.  There are a few improvements that need to be made.  When those are done I feel this will be a very nice addition to the rOpenSci suite of packages.  That it is in my home field is a really nice added bonus.  In short, very nice work.  Now I need to find a reason to dig back into neutral landscape models!

#### Package Review Checklist

Few notes on the Package Review checklist above.  I did not check off Vignettes, Community guidelines and Packaging guidelines.  Details on each:

- Vignettes: There are a number of vignettes and not all are finished and it appears the primary utilitiy for some of them is to populate sections the project website.  Having these unfinished vignettes on CRAN and as part of the installed package should be avoided.  I would like to see just the "getstarted" vignette, and, when completed, the "background" vignette.  I am not sure if this will solve the issue, but perhaps adding the vignettes you don't want pushed to CRAN to `.Rbuildignore` will allow you to build the package for submission but still use those vignettes for the website.  Just a guess on my part that this will work.
  - Also, in "getstarted", I'd drop the raster section.  It is useful, but beyond scope of this vignette.
- Community guidelines:  Main issue here was lack of a CONTRIBUTING or details about contributing in the README.  For an example look at,  <https://github.com/ropensci/scrubr/blob/master/CONTRIBUTING.md>
- Packaging guidelines:  Seem to have most of these done, but a few things still need to be taken care of (e.g. add in the ropensci review badge).  Make sure you adhere to the guidelines at <https://github.com/ropensci/onboarding/blob/master/packaging_guide.md>.  One big one here is the package name.  I know it is already on CRAN as `NLMR`, but changing to `nlmr` would adhere to the guidelines and could be accomplished if you make some changes to the scope of the package.  Details on this I have provided below under "Package Scope."  I'll defer to rOpenSci editors on importance of the lower case package names.

#### Specific Thoughts

See below for some more specific suggestions.

##### Package Scope

Some general thoughts and suggestions on the overal scope of the package.  I personally am a fan of trying to keep packages as streamlined as possible.  `NLMR` currently has two main types of functions, the `nlm_` functions an the `util_` functions.  I could easily see these being split into to separate functions.  If you do this it could faciliate re-naming the package (see above on `NLMR` vs `nlmr`) and creating a new "utilities" package.  This works, I think, because it would make maintaing the pacakges easier, and would also get more exposure to some of the utlity functions which seem a bit more general purpose for rasters and not just neutral landscapes.  Additionally, in [issue 10](https://github.com/marcosci/NLMR/issues/10) you suggest landscape metrics as a possible enhancement.  I personally think incorporating that into `NLMR` would be a mistake since those types of functions have utility well beyond applications geared towards neutral landscapes.  Many of the traditional "FRAGSTATS" type metrics already do exists in the [`SDMTools' package](https://cran.r-project.org/package=SDMTools), but that hasn't seen any updates in ~4 years. As such, developing a standalone landscape metrics package is, I believe, an excellent idea.  In any event, I would encourage (and likely contribute to) a R landscape metrics package, but strongly believe that be done on its own and not incorportated into another pacakge.

##### Dependencies

Somewhat following on my discussion of scope are the number of dependencies that `NLMR` currently has.  I prefer to keep my packages as light as possible and rely on base R whenever possible.  I would like to see some thinning of the dependencies on this.  For example, you import `magrittr`.  While, I love `magrittr` in analysis workflows I avoid its use in packages as it is an unnecessary dependency and, in my experience, can make de-bugging a bit challenging.  Look for opportunities like this to drop some of the packages you import.

##### Future-proofing

It is an exciting time to be working on spatial data in R, but also challenging becuase the underlying packages supporting spatial data in R are in a bit of flux.  Currently, `NLMR` relies on `raster` and `sp`, which are the traditional spatial pacakges.  However, with the release of [`sf`](https://github.com/r-spatial/sf), the development of [`stars`](https://github.com/r-spatial/stars), and discussions about the future of [`raster`](https://github.com/r-spatial/discuss/issues/17) you will need to be actively thinking about how to make sure that `NLMR` stays relevant and incorporates the new directions for spatial data in R.  And as I am writing this, I had the idea that this would be a good opportunity to split up the pacakge (e.g. see Package Scope).  New and more streamlined packages could be developed that also build off of the newer and tidier set of spatial packages.

##### Misc
A few miscellaneous thoughts:

- On documenting the function arguments:  I have not seen the [numerical], [Raster* object] kind of formatting before.  Part of me likes it (i.e. it is precise) but another part doesn't as it is different that most R function documentation.  I'd defer to rOpenSci guidance on this.
- Most of your functions are standardized with arguments for number of rows, columns, extent, rescaling, etc. (i.e. real-world coordinates).  I didn't see these for the curd and whey functions.  Is it possible to have those use the same set of common arguments?  In short, whenever possible, try to make the arguments, and their order, common.
- `util_import_roboto_condensed` is this necessary?  For a package like this it makes more sense to stick with defaults on this so that extra dependencies and functions like this aren't needed.
