### 2018-02-28 #188 @marcosci
thanks @jhollist - I like your thoughts, nice complement to Laura's comments :-)

Will tackle both reviews in the next days and report here.
