### 2018-02-28 #188 @laurajanegraham
Aw thanks @jhollist . I like the suggestion of narrowing the package scope and splitting into two packages. The util functions definitely are useful beyond just NLMs.

I would also be interested in contributing to a standalone landscape metrics package, should this happen.
