### 2018-02-28 #188 @jhollist
Added two small things related to `metric_area` to my review above...  Slipped through the cracks first time around.
