### 2018-03-01 marcosci/NLMR#10 @marcosci
to be continued there: [https://github.com/marcosci/landscapemetrics/issues/1](https://github.com/marcosci/landscapemetrics/issues/1)
