### 2018-03-01 marcosci/NLMR#16 @marcosci
nlm_curds & nlm_wheys
https://github.com/ropensci/onboarding/issues/188#issuecomment-368042693:

No checkmate tests at the beginning.

There is borrowed code between these two. It might make sense to combine them (with a wheys T/F option), or to use the nlm_curds function inside nlm_wheys instead of repeating the code.

These seem quite different in the way they are written to the other functions. Primarily in that the others all specify nrow, ncol, resolution, whereas these specify extent. Is it possible to harmonise these?
