### 2018-03-01 marcosci/NLMR#17 @marcosci
nlm_planargradient 
https://github.com/ropensci/onboarding/issues/188#issuecomment-368042693:

The call to nlm_planargradient (L66) is misspecified. It should be
gradient_raster <- nlm_planargradient(ncol, nrow, direction = direction)

or

gradient_raster <- nlm_planargradient(ncol, nrow, resolution, direction)

or change the order of the arguments to nlm_planargradient
