### 2018-03-01 marcosci/NLMR#18 @marcosci
nlm_fbm
https://github.com/ropensci/onboarding/issues/188#issuecomment-368042693:

nlm_fbm:

Add in checkmate::assert_true(fract_dim > 0).

The code checks that fract_dim is < 2, but the help file states that this parameter is including 2 - fract_dim = numeric in (0, 2]

Values of fract_dim between ~1.56 and ~1.99 generate an error. This is caused by the RandomFields::RFsimulate function.

If this function is run with a random seed, it holds onto it until a new seed is set. This can be fixed by removing the if (!is.null(user_seed)) statement. The default user_seed = NULL will set the seed back to NA.
