### 2018-03-01 marcosci/NLMR#19 @marcosci
nlm_gaussianfield
https://github.com/ropensci/onboarding/issues/188#issuecomment-368042693:

See above issues with setting the random seed in nlm_fBm

Setting the mean value of the field doesn't work because of the line pred_raster <- pred_raster - raster::cellStats(pred_raster, "min"). Is there a purpose for this line of code? Removing this produces Gaussian random field surfaces with the given mean.

Add lower = 0 to the checks for resolution, mag_var and nug

Some additions to the documentation could improve the usability of this function. I would recommend further explanation of what each of the variance parameters do, in a way that the implications for the generated surface can be interpreted. My understanding of these are that:

mag_var sets the variation of the broad landscape
nug sets the variation within the scale of autocorr_range: low nug values will produce a highly spatially autocorrelated surface and high nug values a rougher surface.
Missing a reference, which would be useful for users to get a better understanding of how this function works.
