### 2018-03-01 marcosci/NLMR#20 @marcosci
nlm_mosaic
https://github.com/ropensci/onboarding/issues/188#issuecomment-368042693:

The collect stage overwrites the first step, need to change the for loop on line 91 to for (i in 2:n) (can also remove the i <- 2 from line 86)

In addition, so more information about how these work, and when you would use them would be useful.
