### 2018-03-01 marcosci/NLMR#21 @marcosci
nlm_mpd
https://github.com/ropensci/onboarding/issues/188#issuecomment-368042693:

Generally really clearly explained - more detail on what effect changing rand_dev has on the end landscape would be useful. From some experimenting, it looks like it just changes the overall variance of the landscape
