### 2018-03-01 marcosci/NLMR#24 @marcosci
nlm_randomcluster -> nlm_neigh
https://github.com/ropensci/onboarding/issues/188#issuecomment-368042693:

The neighbourhood input is an integer here, whereas it was method names in nlm_neigh. It would be helpful to unify this (and the spelling of neighbourhood - one US, one British English). My preference would be the numbers, because I always forget the names and it's less to type!
