### 2018-03-01 marcosci/NLMR#25 @marcosci
nlm_randomrectangularcluster
https://github.com/ropensci/onboarding/issues/188#issuecomment-368042693:

Should there also be a test that minl < maxl?
