### 2018-03-01 marcosci/NLMR#26 @marcosci
util_facetplot
https://github.com/ropensci/onboarding/issues/188#issuecomment-368042693:

util_facetplot: is it possible to add a nrow or ncol option into here to control how the facet is laid out?
