### 2018-03-01 ropensci/NLMR#27 @marcosci
Vignettes
https://github.com/ropensci/onboarding/issues/188#issuecomment-369310831:

Vignettes: There are a number of vignettes and not all are finished and it appears the primary utilitiy for some of them is to populate sections the project website. Having these unfinished vignettes on CRAN and as part of the installed package should be avoided. I would like to see just the "getstarted" vignette, and, when completed, the "background" vignette. I am not sure if this will solve the issue, but perhaps adding the vignettes you don't want pushed to CRAN to .Rbuildignore will allow you to build the package for submission but still use those vignettes for the website. Just a guess on my part that this will work.
Also, in "getstarted", I'd drop the raster section. It is useful, but beyond scope of this vignette.

-> finishing background vignette
