### 2018-03-01 marcosci/NLMR#28 @marcosci
Community guidelines
https://github.com/ropensci/onboarding/issues/188#issuecomment-369310831:

Community guidelines: Main issue here was lack of a CONTRIBUTING or details about contributing in the README. For an example look at, https://github.com/ropensci/scrubr/blob/master/CONTRIBUTING.md
