### 2018-03-01 marcosci/NLMR#30 @marcosci
trim dependencies
https://github.com/ropensci/onboarding/issues/188#issuecomment-369310831
