### 2018-03-01 marcosci/NLMR#31 @marcosci
Future-proofing
https://github.com/ropensci/onboarding/issues/188#issuecomment-369310831:

sf + stars ... not sure about stars, but we should exchange the voronoi stuff with the functions from sf
