### 2018-03-01 1ded8b0
fixed function call to nlm_planargradient in nlm_edgegradient. closes #17
