### 2018-03-01 marcosci/NLMR#24 @Z3tt
I would suggest an input called "neigh" or smt similar instead of having two version for both types of spelling.
