### 2018-03-03 #188 @maelle
A few notes:

* It's fine to keep the name all uppercase since it's on CRAN but if you go forward with the splitting changing it to lowercase might be best. By the way if you split the package, the one that'd still be a good fit for onboarding is the non-metrics one.  Recently we had [another submission](https://github.com/ropensci/onboarding/issues/139) where the package ended up being split during review, see [Miles' comment](https://github.com/ropensci/onboarding/issues/139#issuecomment-322713737). If you go forward with splitting, then it might be good to have a chat with MEE editors, since I imagine you'd try to submit a manuscript about two packages born from this review, of which only one would have been transferred. 🤔

* Regarding vignettes that are not on CRAN as inspiration you could have a look at what `usethis` does. There's an article about setup that's on the pkgdown website but not as vignette. https://github.com/r-lib/usethis

Thanks again to both reviewers, and nice work until now @marcosci!
