### 2018-03-04 f678939
fix laura's comments in #19. resolves #19
