+++ man/nlm_gaussianfield.Rd
@@ -19,13 +19,14 @@ Number of rows for the raster.}
 Resolution of the raster.}
 
 \item{autocorr_range}{[\code{numerical(1)}]\cr
-Maximal distance of spatial autocorrelation}
+Maximal distance (raster units) of spatial autocorrelation.}
 
 \item{mag_var}{[\code{numerical(1)}]\cr
-Magnitude of variation in the field}
+Magnitude of variation in the entire landscape.}
 
 \item{nug}{[\code{numerical(1)}]\cr
-Small-scale variations in the field.}
+Magnitude of variation in the scale of \code{autocorr_range},
+a smaller value leads to a more homogeneous landscape.}
 
 \item{mean}{[\code{numerical(1)}]\cr
 Mean value over the field.}
@@ -36,23 +37,30 @@ Set Seed for simulation}
 \item{rescale}{[\code{numeric(1)}]\cr
 If \code{TRUE} (default), the values are rescaled between 0-1.}
 }
-\value{
-RasterLayer
-}
 \description{
 Simulates a spatially correlated random fields (Gaussian random
 fields) model.
 }
+\details{
+Gaussian random fields are a collection of random numbers on a spatially
+discrete set of coordinates (landscape raster). Natural sciences often apply
+them with spatial autocorrelation, meaning that distant things are more
+distinct than closer ones.
+}
 \examples{
 # simulate random gaussian field
 gaussian_field <- nlm_gaussianfield(ncol = 90, nrow = 90,
-                                    autocorr_range = 10,
-                                    mag_var = 3,
-                                    nug = 0.01)
+                                    autocorr_range = 60,
+                                    mag_var = 8,
+                                    nug = 5)
 
 \dontrun{
 # visualize the NLM
 util_plot(gaussian_field)
 }
 
+}
+\references{
+Kéry & Royle (2016) \emph{Applied Hierarachical Modeling in Ecology}
+Chapter 20
 }
