### 2018-03-04 c110414
respond to Laura's comment and elaborate rand_dev. resolves #21
