### 2018-03-04 ef5d4b4
update docu in nlm_randomrectangularcluster. resolves #25
