### 2018-03-05 02c56b4
merge curds and wheys
merge wheys as optional into nlm_curds(). add parameter tests. fix resolution option.

this commit deletes nlm_wheys and corresponding testfile.

resolves #16
