+++ NAMESPACE
@@ -16,7 +16,6 @@ export(nlm_random)
 export(nlm_randomcluster)
 export(nlm_randomelement)
 export(nlm_randomrectangularcluster)
-export(nlm_wheys)
 export(theme_nlm)
 export(theme_nlm_discrete)
 export(theme_nlm_grey)
