### 2018-03-05 0dfb5cb
nlm_polylands split + fix
As recommended by Laura, we turned nlm_polylands into two functions and updated the docu + tests.

resolves #23
