### 2018-03-05 aa1ad98
fix nlm_neigh
Include additional argument checks and get rid of rev.

resolves #22
