+++ R/nlm_curds.R
@@ -5,16 +5,16 @@
 #' @details Random curdling recursively subdivides the plane into blocks.
 #' At each level of the recursion, a fraction of the this block is declared as
 #' habitat (value == TRUE) while the remaining stays matrix (value == FALSE).
-#' 
-#' With the optional argument for wheys, previously selected cells that were 
-#' declared matrix (value == FALSE) during recursion, can now contain a 
+#'
+#' With the optional argument for wheys, previously selected cells that were
+#' declared matrix (value == FALSE) during recursion, can now contain a
 #' proportion (\code{q}) of habitat cells.
-#' 
+#'
 #' If \deqn{p_{1} = p_{2} = q_{2} = ... = p_{n} = p_{n}} the models resembles
 #' a binary random map.
-#' 
+#'
 #' Note that you can not set ncol and nrow with this landscape algorithm.
-#' The amount of cells is given by the vectorproduct of s.
+#' The amount of cells and hence dimension of the raster is given by the vectorproduct of s.
 #'
 #' @param p [\code{numerical(x)}]\cr
 #' Vector with percentage(s) to fill with curds (fill with Habitat (value ==
@@ -23,7 +23,6 @@
 #' Vector of successive cutting steps for the blocks (split 1 block into x
 #' blocks).
 #' @param q [\code{numerical(x)}]\cr
-#' Vector of with percentage(s) to fill with wheys (fill with Habitat (value ==
 #' TRUE)).
 #' @param resolution [\code{numerical(1)}]\cr
 #' Resolution of the resulting raster.
