### 2018-03-05 054055b
neighbourhood spelling
unified neighbourhood spelling everwhere and argument neighbourhood in nlm_neigh now takes integers instead of characters

resolves #22
