+++ tests/testthat/test_neigh.R
@@ -6,7 +6,7 @@ neigh_raster  <- nlm_neigh(ncol = 20,
                            p_neigh = 0.1,
                            p_empty = 0.3,
                            categories = 5,
-                           neighborhood = 4
+                           neighbourhood = 4
                            )
 
 test_that("nlm_neigh behaves like it should", {
@@ -29,7 +29,7 @@ neigh_raster  <- nlm_neigh(ncol = 20,
                            p_neigh = 0.1,
                            p_empty = 0.3,
                            categories = 5,
-                           neighborhood = 8
+                           neighbourhood = 8
 )
 
 test_that("nlm_neigh behaves like it should", {
