### 2018-03-06 fd2acd7
some lint fixes and get rid of magrittr in nlm_curds
