+++ R/nlm_curds.R
@@ -34,7 +34,7 @@
 #' # simulate random curdling
 #' random_curdling <- nlm_curds(p = c(0.5, 0.3, 0.6), s = c(32, 6, 2))
 #' # simulate wheyed curdling
-#' wheyed_curdling <- nlm_curds(p = c(0.5, 0.3, 0.6), s = c(32, 6, 2), q = c(0.6, 0.05, 0.2))
+#' wheyed_curdling <- nlm_curds(p = c(0.5, 0.3, 0.6), s = c(32, 6, 2), q = c(0.1, 0.05, 0.2))
 #' \dontrun{
 #' # Visualize the NLMs
 #' util_plot(random_curdling, discrete = TRUE)
@@ -91,7 +91,7 @@ nlm_curds <- function(p,
 
     # 'wheying' select ids randomly which are to be set to false and do so
     if (!is.null(q)) {
-      vl_tib_whey           <- dplyr::filter(vl_tib, !value)
+      vl_tib_whey           <- dplyr::filter(vl_tib, value)
       vl_tib_whey           <- dplyr::sample_frac(vl_tib_whey, q[i])
       vl_tib_whey           <- vl_tib_whey$id
       vl_tib$value[vl_tib_whey] <- FALSE
