### 2018-03-06 d088e6e
include community guidelines
rearranged readme and included CONTRIBUTING.md

resolves #28
