+++ CONTRIBUTING.md
@@ -1,0 +1,24 @@
+# CONTRIBUTING #
+
+### Please contribute!
+
+We love collaboration.
+
+### Bugs?
+
+* Submit an issue on the Issues page [here](https://github.com/marcosci/NLMR/issues)
+
+### Code contributions
+
+* Fork this repo to your Github account
+* Clone your version on your account down to your machine from your account, e.g,. `git clone https://github.com/<yourgithubusername>/NLMR.git`
+* Make sure to track progress upstream (i.e., on our version of `scrubr` at `ropensci/scrubr`) by doing `git remote add upstream https://github.com/ropensci/NLMR.git`. Before making changes make sure to pull changes in from upstream by doing either `git fetch upstream` then merge later or `git pull upstream` to fetch and merge in one step
+* Make your changes (bonus points for making changes on a new branch)
+* If you alter package functionality at all (e.g., the code itself, not just documentation)
+please do write some tests to cover the new functionality.
+* Push up to your account
+* Submit a pull request to home base at `ropensci/NLMR`
+
+### Questions? Get in touch: [sciaini.marco@gmail.com](mailto:sciaini.marco@gmail.com)
+
+### Thanks for contributing!
