+++ README.Rmd
@@ -80,21 +80,15 @@ raster::as.matrix(nlm_raster) %>%
 
 ```
 
+## Meta
 
-
-## Citation
-
-```{r, echo  = FALSE, result = "asis", comment = ""}
-citation("NLMR")
-```
-
-Additionally, we keep a [record of publications](https://marcosci.github.io/NLMR/articles/publication_record.html/) that use`NLMR`.
-Hence, if you used `NLMR` please [file an issue on GitHub](https://github.com/marcosci/NLMR/issues/new/) so we can add it to the list.
-
-## Contributor Code of Conduct
-
-Please note that this project is released with a [Contributor Code of Conduct](CONDUCT.md). By participating in this project you agree to abide by its terms.
-
+* Please [report any issues or bugs](https://github.com/marcosci/NLMR/issues/new/).
+* License: GPL3
+* Get citation information for `NLMR` in R doing `citation(package = 'NLMR')`
+   * Additionally, we keep a [record of publications](https://marcosci.github.io/NLMR/articles/publication_record.html/) that use`NLMR`. Hence, if you used `NLMR` please [file an issue on GitHub](https://github.com/marcosci/NLMR/issues/new/) so we can add it to the list.
+* We are very open to contributions - if you are interested check [Contributor Code of Conduct](CONTRIBUTING.md).
+   * Please note that this project is released with a [Contributor Code of Conduct](CONDUCT.md). By participating in this project you agree to abide by its terms.
+   
 ## Dependencies
 
 `NLMR` imports many great packages that it depends on. Many thanks to the developers of these tools:
