### 2018-03-06 60888a8
facetplot layout
add nrow, ncol options to facetplot and pass arguments to facetwrap.

resolves #26
