### 2018-03-07 9c63762
introduce sf and fasterize in mosaictess
... and retire nlm_randomelement, as it is now obsolete.
