+++ man/nlm_randomcluster.Rd
@@ -39,7 +39,7 @@ The implemented algorithm has been adopted from Etherington et al. 2014 and is i
 }
 \examples{
 # simulate random clustering
-random_cluster <- nlm_randomcluster(ncol = 20, nrow = 10, resolution = 1,
+random_cluster <- nlm_randomcluster(ncol = 20, nrow = 10,
                                     neighbourhood = 4, p = 0.4)
 \dontrun{
 # visualize the NLM
