+++ NAMESPACE
@@ -15,7 +15,6 @@ export(nlm_percolation)
 export(nlm_planargradient)
 export(nlm_random)
 export(nlm_randomcluster)
-export(nlm_randomelement)
 export(nlm_randomrectangularcluster)
 export(theme_nlm)
 export(theme_nlm_discrete)
