### 2018-03-08 marcosci/NLMR#33 @marcosci
nlm_randomcluster 
skip nlmpy and use original santiago saura algorithm
