### 2018-03-08 2e6abcd
add value check and automated random seed reset
