+++ R/nlm_fBm.R
@@ -54,7 +54,8 @@ nlm_fBm <- function(ncol,
   checkmate::assert_count(nrow, positive = TRUE)
   checkmate::assert_numeric(resolution)
   checkmate::assert_numeric(fract_dim)
-  checkmate::assert_true(fract_dim < 2)
+  checkmate::assert_true(fract_dim > 0)
+  checkmate::assert_true(fract_dim <= 2)
   checkmate::assert_logical(rescale)
 
   # specify RandomFields options ----
@@ -62,18 +63,16 @@ nlm_fBm <- function(ncol,
   RandomFields::RFoptions(spConform = FALSE)
 
   # set RF seed ----
-  if (!is.null(user_seed)) {
-    RandomFields::RFoptions(seed = user_seed)
-  }
+  RandomFields::RFoptions(seed = user_seed)
 
   # formulate and simulate fBm model
   fbm_model <- RandomFields::RMfbm(
     alpha = fract_dim)
   fbm_simu <- RandomFields::RFsimulate(fbm_model,
                                        # fBm changes x and y?
-                                       y = seq(ncol),
-                                       x = seq(nrow),
-                                       grid =  TRUE)
+                                       y = seq.int(0, length.out = ncol),
+                                       x = seq.int(0, length.out = nrow),
+                                       grid = TRUE)
 
 
   # transform simulation into raster ----
