### 2018-03-08 marcosci/NLMR#18 @mattfrit
opened fix/18 branch. added check and seed reset.

the error with the alpha values is somewhere in rfInit. it is possible to produce output by switching modus_operandi to "sloppy"...
