### 2018-03-08 5f6158e
hurray, we're sp free ...
refactored everything sp related to sf.
rasterization now with fasterize.

resolves #31
