### 2018-03-08 marcosci/NLMR#20 @marcosci
closed in the process of issue #31
