### 2018-03-08 af5cc51
update ci
include fasterize
