+++ appveyor.yml
@@ -16,6 +16,7 @@ cache:
 # Adapt as necessary starting from here
 
 build_script:
+  - travis-tool.sh install_github ecohealthalliance/fasterize
   - travis-tool.sh install_deps
 
 test_script:
