### 2018-03-09 af5f7d9
update travis
... still no clue why appveyor fails to build
