+++ .travis.yml
@@ -10,6 +10,12 @@ r_github_packages:
   - r-lib/covr
   - ecohealthalliance/fasterize
 
+before_install:
+  - sudo add-apt-repository ppa:ubuntugis/ubuntugis-unstable --yes
+  - sudo apt-get --yes --force-yes update -qq
+  - sudo apt-get install --yes libudunits2-dev libproj-dev libgeos-dev libgdal-dev
+  - Rscript -e 'update.packages(ask = FALSE)'
+
 after_success:
   - Rscript -e 'covr::codecov(branch = "develop")'
 
