### 2018-03-09 a483a37
include option modus_operandi
... can be set to sloppy, to allow  for fractal dimensions between 1.6 and 1.9

resolves #18
