+++ appveyor.yml
@@ -14,6 +14,8 @@ cache:
   - C:\RLibrary
 
 # Adapt as necessary starting from here
+environment:
+  - USE_RTOOLS: true
 
 build_script:
   - travis-tool.sh install_github ecohealthalliance/fasterize
