### 2018-03-14 1dcca3e
nlm_randomcluster enhancement
rewrite randomcluster algorithm. it is now a direct implementation of steps described in Saura (2000).

resolves #33
