+++ NAMESPACE
@@ -32,5 +32,4 @@ export(util_raster2tibble)
 export(util_rescale)
 export(util_tibble2raster)
 export(util_w2cp)
-importFrom(igraph,components)
 importFrom(magrittr,"%>%")
