### 2018-03-14 f80c1f4
start refactoring in nlmr and landscapetools
