+++ R/theme_nlm.R
@@ -1,464 +1,0 @@
-#' theme_nlm
-#'
-#' @description Opinionated ggplot2 theme to visualize NLM raster.
-#'
-#' @param base_family [\code{character}()]  base font family size
-#' @param base_size [\code{numeric}(1)] base font size
-#' @param plot_title_family [\code{character}()] plot title family
-#' @param plot_title_face [\code{character}()] plot title face
-#' @param plot_title_size [\code{character}()] plot title size
-#' @param plot_title_margin [\code{numeric}(1)] plot title margin
-#' @param plot_margin [\code{character}()]
-#' plot margin (specify with \code{ggplot2::margin})
-#' @param subtitle_family [\code{character}()] plot subtitle family
-#' @param subtitle_face [\code{character}()]` plot subtitle face
-#' @param subtitle_size [\code{numeric}(1)] plot subtitle size
-#' @param subtitle_margin [\code{numeric}(1)]
-#' plot subtitle margin bottom (single numeric value)
-#' @param strip_text_family [\code{character}()] facet facet label font family
-#' @param strip_text_face [\code{character}()] facet facet label font face
-#' @param strip_text_size [\code{numeric}(1)]
-#' facet label font family, face and size
-#' @param caption_family [\code{character}()] plot caption family
-#' @param caption_face [\code{character}()] plot caption face
-#' @param caption_size [\code{numeric}(1)] plot caption size
-#' @param caption_margin [\code{numeric}(1)] plot caption margin
-#' @param legend_title [\code{character}()] Title of the legend (default "Z")
-#' @param legend_labels [\code{character}()] Labels for the legend ticks, if
-#' used with \code{\link{util_plot}} they are automatically derived.
-#' @param ratio [\code{character}()]
-#' ratio for tiles (default 1, if your raster is not a square the ratio should
-#' be \code{raster::nrow(x) / raster::ncol(x)})
-#' @param viridis_scale [\code{character}(1)]
-#' Five options are available: "viridis - magma" (= "A"),
-#'                             "viridis - inferno" (= "B"),
-#'                             "viridis - plasma" (= "C"),
-#'                             "viridis - viridis" (= "D",  the default option)
-#'
-#' @details
-#' A focused theme to visualize raster data that sets a lot of defaults for the
-#' \code{ggplot2::theme}.
-#'
-#' The theme can make use of the Roboto Condensed font (Open Source font from
-#' Google).
-#' If your local font library does not contain Roboto as a font, you can
-#' import it via \code{\link{util_import_roboto_condensed}} (highly recommended).
-#'
-#' The functions are setup in such a way that you can customize your own one by
-#' just wrapping the call and changing the parameters.
-#' The theme itself is heavily influenced by hrbrmstr and his package
-#' hrbrthemes (\url{https://github.com/hrbrmstr/hrbrthemes/}).
-#'
-#' @seealso \code{\link{util_import_roboto_condensed}}
-#'
-#' @examples
-#' # nolint start
-#' \dontrun{
-#' # simulate NLM
-#' x <- nlm_random(ncol = 75,
-#'                 nrow = 75)
-#' # classify
-#' y <- c(0.5, 0.15, 0.25)
-#' y <- util_classify(x, y, c("1", "2", "3"))
-#'
-#' # color + continuous
-#' rasterVis::gplot(x) +
-#'   ggplot2::geom_tile(ggplot2::aes(fill = value)) +
-#'   ggplot2::labs(x = "Easting",
-#'                 y = "Northing") +
-#'   theme_nlm() +
-#'   ggplot2::ggtitle("Random NLM with continuous viridis color scale",
-#'                    subtitle = "75x75 cells") +
[…]
