+++ vignettes/background.Rmd
@@ -20,7 +20,7 @@ Neutral landscape models, originally developed as a null model against which to 
 
 A series of such neutral landscape patterns can be simulated in which concerned variables are changed while other variables are kept constant. With this approach, the influence of the concerned variable on the object of study can be determined. In addition to empirical data, the simulation of landscapes with extreme configurations is possible, thus allowing us to find thresholds for conservation related landscape processes such as fragmentation. 
 
-So far, most of the software available to generate neutral landscapes only perform a limited set of algorithms or are implemented in programming languages that are not widely used in the field of computational ecology. Our R package, NLMR, aims to be a broad library of neutral landscape models that can easily extend in existing landscape analyses. NLMR combines a large number of algorithms from published software for simulating neutral landscapes and includes utility functions to classify and combine the landscapes. The simulation results are obtained in a geospatial data format (raster* objects from the raster package) and can, therefore, be used in any sort of raster data operation that is performed with standard observation data. Furthermore, the data format allows ecologists to use their simulations in most GIS software programs allowing for easy communication of their research.
+So far, most of the software available to generate neutral landscapes only perform a limited set of algorithms or are implemented in programming languages that are not widely used in the field of computational ecology. Our R package, nlmr, aims to be a broad library of neutral landscape models that can easily extend in existing landscape analyses. nlmr combines a large number of algorithms from published software for simulating neutral landscapes and includes utility functions to classify and combine the landscapes. The simulation results are obtained in a geospatial data format (raster* objects from the raster package) and can, therefore, be used in any sort of raster data operation that is performed with standard observation data. Furthermore, the data format allows ecologists to use their simulations in most GIS software programs allowing for easy communication of their research.
 
 ### Concept
 
