+++ vignettes/publication_record.Rmd
@@ -6,4 +6,4 @@ vignette: >
   %\VignetteEncoding{UTF-8}
 ---
 
-We try to maintain a list here of publications that have used the `NLMR` package. If you have used `NLMR` in your own work, we would love to hear from you. Please [file an issue on GitHub](https://github.com/marcosci/NLMR/issues/new/) so we can add your work to this list.
+We try to maintain a list here of publications that have used the `nlmr` package. If you have used `nlmr` in your own work, we would love to hear from you. Please [file an issue on GitHub](https://github.com/marcosci/nlmr/issues/new/) so we can add your work to this list.
