+++ man/nlmr-package.Rd
@@ -1,33 +1,33 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/NLMR.R
 \docType{package}
-\name{NLMR-package}
-\alias{NLMR}
-\alias{NLMR-package}
+\name{nlmr-package}
+\alias{nlmr}
+\alias{nlmr-package}
 \title{Simulating Neutral Landscape Models}
 \description{
-\emph{NLMR} is an R package for simulating neutral landscape models (NLM).
+\emph{nlmr} is an R package for simulating neutral landscape models (NLM).
 }
 \details{
 This package contains vignettes that introduce NLM and basic usage
-of the \emph{NLMR} package. The vignettes in this package are listed below.
+of the \emph{nlmr} package. The vignettes in this package are listed below.
 
 \describe{
-\item{\href{https://marcosci.github.io/NLMR/articles/background.html/}{NLM
+\item{\href{https://marcosci.github.io/nlmr/articles/background.html/}{NLM
 Concepts + Terminology}}{Background information on the concepts and
 terminology that underpin the simulation of neutral landscape models.}
-\item{\href{https://marcosci.github.io/NLMR/articles/bestiary.html/}{NLMR
+\item{\href{https://marcosci.github.io/nlmr/articles/bestiary.html/}{nlmr
 Bestiary}}{Bestiary of the available neutral landscape models.}
-\item{\href{https://marcosci.github.io/NLMR/articles/getstarted.html/}{
-Quickstart Guide}}{Short walk-through of the \emph{NLMR} package and how to
+\item{\href{https://marcosci.github.io/nlmr/articles/getstarted.html/}{
+Quickstart Guide}}{Short walk-through of the \emph{nlmr} package and how to
 handle the simulations.}
 }
 }
 \seealso{
 Useful links:
 \itemize{
-  \item \url{https://marcosci.github.io/NLMR/}
-  \item Report bugs at \url{https://github.com/marcosci/NLMR/issues/}
+  \item \url{https://marcosci.github.io/nlmr/}
+  \item Report bugs at \url{https://github.com/marcosci/nlmr/issues/}
 }
 
 }

[…]
