+++ README.Rmd
@@ -2,6 +2,7 @@
 output:
   github_document:
     html_preview: false
+bibliography: references.bib
 ---
 
 
@@ -18,9 +19,9 @@ knitr::opts_chunk$set(
 [![Build Status](https://travis-ci.org/marcosci/nlmr.svg?branch=develop)](https://travis-ci.org/marcosci/nlmr)
 [![Build status](https://ci.appveyor.com/api/projects/status/ns75pdrbaykxc865?svg=true)](https://ci.appveyor.com/project/marcosci/nlmr)
 [![codecov](https://codecov.io/gh/marcosci/nlmr/branch/develop/graph/badge.svg?token=MKCm2fVrDa)](https://codecov.io/gh/marcosci/nlmr)
-[![CRAN\_Status\_Badge](http://www.r-pkg.org/badges/version/nlmr)](https://cran.r-project.org/package=nlmr) 
+[![CRAN\_Status\_Badge](http://www.r-pkg.org/badges/version/nlmr)](https://cran.r-project.org/package=NLMR) 
 [![Join the chat at  https://gitter.im/nlmr_landscapegenerator](https://badges.gitter.im/Join%20Chat.svg)](https://gitter.im/nlmr_landscapegenerator) 
-[![](http://cranlogs.r-pkg.org/badges/grand-total/nlmr)](http://cran.rstudio.com/web/packages/nlmr/index.html) 
+[![](http://cranlogs.r-pkg.org/badges/grand-total/nlmr)](http://cran.rstudio.com/web/packages/NLMR/index.html) 
 [![](https://badges.ropensci.org/188_status.svg)](https://github.com/ropensci/onboarding/issues/188)
 
 # nlmr <img src="vignettes/logo.png" align="right"  height="175" />
@@ -38,8 +39,10 @@ Furthermore, it simulates NLMs within a self-contained, reproducible framework.
 Install the release version from CRAN:
 
 ```{r eval = FALSE}
-install.packages("nlmr")
+install.packages("NLMR")
 ```
+
+__NOTE__: The version on CRAN is 
 
 To install the developmental version of ``nlmr``, use the following R code:
 
@@ -50,34 +53,109 @@ devtools::install_github("marcosci/nlmr", ref = "develop")
 
 ## Example
 
-Here we will provide a simple example on using ``nlmr``:
+Simulating a neutral landscape model with `nlmr` can be achieved in a single line:
 
-```{r message=FALSE, fig.height=7, fig.width=7, fig.align='center'}
-library(nlmr)
+```{r message=FALSE}
+# Simulate 50x50 random cluster raster
+nlm_raster <- nlmr::nlm_randomcluster(nrow = 50,
+                                      ncol = 50,
+                                      p    = 0.5,
+                                      ai   = c(0.2, 0.2, 0.5, 0.1))
+```
+
+Landscape models simulated by nlmr are returned as raster objects, hence they can
+be processed as usual, e.g. visualized:
+
+```{r, fig.height=7, fig.width=7, fig.align='center'}
+rasterVis::levelplot(nlm_raster, margin = FALSE, par.settings = rasterVis::viridisTheme())
+```
+
+## Overview
+
+`nlmr` supplies 14 NLM algorithms, with several options to simulate derivates of
+them. The algorithms differ from each other in spatial auto-correlation, from 
+no auto-correlation (random NLM) to a constant gradient (planar gradients):  
+
+```{r warning=FALSE, message = FALSE, results='asis', echo=FALSE, cache=FALSE}
+library(tibble)
 library(magrittr)
-library(ggplot2)  # to extend the plot functionality of nlmr 
-library(SDMTools) # to calculate basic landscape metrics
+library(knitr)
+library(kableExtra)
 
-# Simulate 50x50 rectangular cluster raster
[…]
