+++ README.md
@@ -1,6 +1,6 @@
 
 <!-- README.md is generated from README.Rmd. Please edit that file -->
-[![Build Status](https://travis-ci.org/marcosci/nlmr.svg?branch=develop)](https://travis-ci.org/marcosci/nlmr) [![Build status](https://ci.appveyor.com/api/projects/status/ns75pdrbaykxc865?svg=true)](https://ci.appveyor.com/project/marcosci/nlmr) [![codecov](https://codecov.io/gh/marcosci/nlmr/branch/develop/graph/badge.svg?token=MKCm2fVrDa)](https://codecov.io/gh/marcosci/nlmr) [![CRAN\_Status\_Badge](http://www.r-pkg.org/badges/version/nlmr)](https://cran.r-project.org/package=nlmr) [![Join the chat at https://gitter.im/nlmr\_landscapegenerator](https://badges.gitter.im/Join%20Chat.svg)](https://gitter.im/nlmr_landscapegenerator) [![](http://cranlogs.r-pkg.org/badges/grand-total/nlmr)](http://cran.rstudio.com/web/packages/nlmr/index.html) [![](https://badges.ropensci.org/188_status.svg)](https://github.com/ropensci/onboarding/issues/188)
+[![Build Status](https://travis-ci.org/marcosci/nlmr.svg?branch=develop)](https://travis-ci.org/marcosci/nlmr) [![Build status](https://ci.appveyor.com/api/projects/status/ns75pdrbaykxc865?svg=true)](https://ci.appveyor.com/project/marcosci/nlmr) [![codecov](https://codecov.io/gh/marcosci/nlmr/branch/develop/graph/badge.svg?token=MKCm2fVrDa)](https://codecov.io/gh/marcosci/nlmr) [![CRAN\_Status\_Badge](http://www.r-pkg.org/badges/version/nlmr)](https://cran.r-project.org/package=NLMR) [![Join the chat at https://gitter.im/nlmr\_landscapegenerator](https://badges.gitter.im/Join%20Chat.svg)](https://gitter.im/nlmr_landscapegenerator) [![](http://cranlogs.r-pkg.org/badges/grand-total/nlmr)](http://cran.rstudio.com/web/packages/NLMR/index.html) [![](https://badges.ropensci.org/188_status.svg)](https://github.com/ropensci/onboarding/issues/188)
 
 nlmr <img src="vignettes/logo.png" align="right"  height="175" />
 =================================================================
@@ -17,8 +17,10 @@ Installation
 Install the release version from CRAN:
 
 ``` r
-install.packages("nlmr")
+install.packages("NLMR")
 ```
+
+**NOTE**: The version on CRAN is
 
 To install the developmental version of `nlmr`, use the following R code:
 
@@ -30,52 +32,45 @@ devtools::install_github("marcosci/nlmr", ref = "develop")
 Example
 -------
 
-Here we will provide a simple example on using `nlmr`:
+Simulating a neutral landscape model with `nlmr` can be achieved in a single line:
 
 ``` r
-library(nlmr)
-library(magrittr)
-library(ggplot2)  # to extend the plot functionality of nlmr 
-library(SDMTools) # to calculate basic landscape metrics
-
-# Simulate 50x50 rectangular cluster raster
-nlm_raster <- nlm_randomrectangularcluster(50,50, resolution = 1, minl = 3, maxl = 7)
-
-# Plot the NLM
-util_plot(nlm_raster) +
-  labs(title="Random rectangular cluster NLM \n (50x50 cells)")
-#> Warning in seq.default(.limits[1], .limits[2], length = guide$nbin):
-#> partial argument match of 'length' to 'length.out'
+# Simulate 50x50 random cluster raster
+nlm_raster <- nlmr::nlm_randomcluster(nrow = 50,
[…]
