### 2018-03-15 817e5e2
get rid of utility functions
