+++ R/nlm_distancegradient.R
@@ -29,7 +29,7 @@
 #'                                            origin = c(20, 30, 10, 15)))
 #' \dontrun{
 #' # visualize the NLM
-#' util_plot(distance_gradient)
+#' rasterVis::levelplot(distance_gradient, margin = FALSE, par.settings = rasterVis::viridisTheme())
 #' }
 #' @seealso \code{\link{nlm_edgegradient}},
 #' \code{\link{nlm_planargradient}}
