+++ R/nlm_randomrectangularcluster.R
@@ -27,13 +27,14 @@
 #'
 #' @examples
 #' # simulate random rectangular cluster
-#' randomrectangular_cluster <- nlm_randomrectangularcluster(ncol = 30,
+#' (randomrectangular_cluster <- nlm_randomrectangularcluster(ncol = 30,
 #'                                                           nrow = 30,
 #'                                                           minl = 5,
-#'                                                           maxl = 10)
+#'                                                           maxl = 10))
 #' \dontrun{
 #' # visualize the NLM
-#' util_plot(randomrectangular_cluster)
+#' rasterVis::levelplot(randomrectangular_cluster, margin = FALSE,
+#' par.settings = rasterVis::viridisTheme())
 #' }
 #'
 #' @aliases nlm_randomrectangularcluster
