+++ R/util_rescale.R
@@ -9,7 +9,6 @@
 #' @details Rasters generated by \code{nlm_} functions are scaled between 0 and 1 as default, this option can be set to \code{FALSE} if needed.
 #'
 #' @keywords internal
-#' @export
 #'
 
 util_rescale <- function(x) {
