+++ R/nlm_mosaictess.R
@@ -23,11 +23,11 @@
 #'
 #' @examples
 #' # simulate polygonal landscapes
-#' mosaictess <- nlm_mosaictess(ncol = 30, nrow = 60, germs = 200)
+#' (mosaictess <- nlm_mosaictess(ncol = 30, nrow = 60, germs = 200))
 #'
 #' \dontrun{
 #' # visualize the NLM
-#' util_plot(mosaictess)
+#' rasterVis::levelplot(mosaictess, margin = FALSE, par.settings = rasterVis::viridisTheme())
 #' }
 #'
 #' @references
