### 2018-03-15 4e26eba
some fixes after running goodpractice
