+++ references.bib
@@ -8,8 +8,7 @@
   pages    = {164--168},
   issn     = {2041-210X},
   doi      = {10.1111/2041-210X.12308},
-  keywords = {artificial landscape, geographic information system, GIS, landscape generator, NLM, NumPy, SciPy, virtual landscape},
-  url      = {http://dx.doi.org/10.1111/2041-210X.12308},
+  keywords = {artificial landscape, geographic information system, GIS, landscape generator, NLM, NumPy, SciPy, virtual landscape}
 }
 
 @Article{Keitt2000,
