### 2018-03-16 70bee6f
exchange fasterize with velox
