+++ R/nlm_mosaicgibbs.R
@@ -93,9 +93,15 @@ nlm_mosaicgibbs <- function(ncol,
     )
 
   # (f)rasterize with lightning speed ----
-  r <- raster::raster(raster::extent(c(0, ncol, 0, nrow)), res = resolution)
-  r <- fasterize::fasterize(voronoi_tess, r, field = "value", fun = "sum")
+  vx <- velox::velox(matrix(0, nrow = nrow, ncol = ncol),
+                     extent=raster::extent(voronoi_tess),
+                     res = c(resolution,resolution))
+  vx$rasterize(voronoi_tess, field = "value")
+  r <- vx$as.RasterLayer()
 
+  # return to fasterize when it is on CRAN
+  # r <- raster::raster(raster::extent(voronoi_tess), res = resolution)
+  # r <- fasterize::fasterize(voronoi_tess, r, field = "value", fun = "sum")
 
   # specify resolution ----
   raster::extent(r) <- c(0,
