+++ DESCRIPTION
@@ -35,7 +35,7 @@ Imports:
     spatstat,
     stats,
     tibble,
-    fasterize,
+    velox,
     sf
 URL: https://marcosci.github.io/nlmr/
 BugReports: https://github.com/marcosci/nlmr/issues/
@@ -49,4 +49,3 @@ Suggests:
     rasterVis
 VignetteBuilder:
     knitr
-Remotes: ecohealthalliance/fasterize
