+++ references.bib
@@ -7,7 +7,6 @@
   number   = {2},
   pages    = {164--168},
   issn     = {2041-210X},
-  doi      = {10.1111/2041-210X.12308},
   keywords = {artificial landscape, geographic information system, GIS, landscape generator, NLM, NumPy, SciPy, virtual landscape}
 }
 
