+++ R/nlm_mosaictess.R
@@ -72,7 +72,7 @@ nlm_mosaictess <- function(ncol,
 
   # (f)rasterize with lightning speed ----
   vx <- velox::velox(matrix(0, nrow = nrow, ncol = ncol),
-                     extent=raster::extent(voronoi_tess),
+                     extent = raster::extent(voronoi_tess),
                      res = c(resolution,resolution))
   vx$rasterize(voronoi_tess, field = "value")
   r <- vx$as.RasterLayer()
