### 2018-03-16 3b2e808
center figure in readme
