### 2018-03-19 0372778
change some last NLMR remnants
