+++ man/nlmr-package.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/NLMR.R
+% Please edit documentation in R/nlmr.R
 \docType{package}
 \name{nlmr-package}
 \alias{nlmr}
@@ -37,12 +37,16 @@ Useful links:
 Authors:
 \itemize{
   \item Matthias Fritsch \email{matthias.fritsch@forst.uni-goettingen.de}
-  \item Craig Simpkins \email{simpkinscraig063@gmail.com}
 }
 
 Other contributors:
 \itemize{
-  \item Cédric Scherer \email{cedricphilippscherer@gmail.com} (Implemented nlm_neigh) [contributor]
+  \item Craig Simpkins \email{simpkinscraig063@gmail.com} (Craig reviewed/provided comments to improve the documentation.) [contributor]
+  \item Cédric Scherer \email{cedricphilippscherer@gmail.com} (Cédric implemented nlm_neigh) [contributor]
+  \item Laura Graham (Laura reviewed the package for rOpenSci, see 
+                  https://github.com/ropensci/onboarding/issues/188) [NA]
+  \item Jeffrey Hollister (Jeffrey reviewed the package for rOpenSci, see 
+                  https://github.com/ropensci/onboarding/issues/188) [NA]
 }
 
 }
