+++ DESCRIPTION
@@ -3,12 +3,21 @@ Type: Package
 Title: Simulating Neutral Landscape Models
 Version: 0.2.1
 Authors@R: c(person("Marco", "Sciaini", email = "sciaini.marco@gmail.com", role = c("aut", "cre")), 
-    person("Matthias", "Fritsch", email = "matthias.fritsch@forst.uni-goettingen.de", role =  c("aut")),
-    person("Craig", "Simpkins", email = "simpkinscraig063@gmail.com", role =  c("aut")),
-    person("Cédric", "Scherer", email = "cedricphilippscherer@gmail.com", role =  c("ctb"), comment = "Implemented nlm_neigh"))
+    person("Matthias", "Fritsch", email = "matthias.fritsch@forst.uni-goettingen.de", role = "aut"),
+    person("Craig", "Simpkins", email = "simpkinscraig063@gmail.com",
+       role =  "ctb", comment = "Craig reviewed/provided comments to improve the documentation."),
+    person("Cédric", "Scherer", email = "cedricphilippscherer@gmail.com", 
+       role =  c("ctb"), comment = "Cédric implemented nlm_neigh"), 
+    person("Laura", "Graham", role = "rev",
+       comment = "Laura reviewed the package for rOpenSci, see 
+                  https://github.com/ropensci/onboarding/issues/188"),
+    person("Jeffrey", "Hollister", role = "rev",
+       comment = "Jeffrey reviewed the package for rOpenSci, see 
+                  https://github.com/ropensci/onboarding/issues/188"))
 Maintainer: Marco Sciaini <sciaini.marco@gmail.com>
-Description: Provides neutral landscape models (<https://doi.org/10.1007/BF02275262>, <http://sci-hub.tw/10.1007/bf02275262>)
-    that can easily extend in existing landscape analyses. Neutral landscape models range from "hard" 
+Description: Provides neutral landscape models (<https://doi.org/10.1007/BF02275262>,
+    <http://sci-hub.tw/10.1007/bf02275262>).  
+    Neutral landscape models range from "hard" 
     neutral models (completely random distributed), to "soft" neutral models 
     (definable spatial characteristics) and generate landscape patterns that are 
     independent of ecological processes.
