### 2018-03-20 #188 @marcosci
## Response to reviews

@laurajanegraham and @jeffhollist - thanks again for your reviews

We decided to follow the recommendations of Laura and Jeff to split the package. This leaves us with:

- [nlmr](https://github.com/marcosci/nlmr)
  - nlmr is now a trimmed down version of NLMR, only containing the algorithms to simulate neutral landscape models. Hence, the (direct) dependencies are quite small.
- [landscapetools](https://github.com/marcosci/landscapetools)
  - Here we gathered all the utility functions and visualization functions from NLMR. There is no package to facilitate the workflow in the context of landscape analysis, so there should be potential for growth.
- [landscapemetrics](https://github.com/marcosci/landscapemetrics)
  - Since there wasn't much implemented in NLMR yet, it is still pretty empty. But Laura said that she would be interested in working on it with us and Jeff will be our mascot - looking forward to it :)

@maelle, you said:
> By the way if you split the package, the one that'd still be a good fit for onboarding is the non-metrics one.

... my gut feeling says that only the nlmr package is now of interest and not the combination of landscapetools and nlmr (however this combination may look like - meta package, second onboarding issue, ...)? If I am mistaken, I will provide comments on the reviewers' remarks on the utility functions. Otherwise, I will focus for now on nlmr.

### responses to @laurajanegraham comments
> nlm_curds & nlm_wheys:

We combined both functions. You get the wheyed pattern now by providing the argument `wheyes` ([#16](https://github.com/marcosci/nlmr/issues/16)). Since you recursively subdivide the landscape into smaller and smaller blocks, I can not think of an efficient way to implement this whole specifically controlling for nrow and ncol.

> nlm_edgegradient:

We did that here [#17](https://github.com/marcosci/nlmr/issues/17).

> nlm_fbm:

Values of fract_dim between ~1.56 and ~1.99 should work now if you set the RandomFields option
to sloppy. You can do that now via the `nlm_fBm` function - we are not quite sure why you have to do
this, but we contacted the maintainer of RandomFields and as soon as we know more, we will implement it.
At least the functions produce now valid patterns for the whole range of fract_dim.
We also implemented the rest of your comments, see
[#18](https://github.com/marcosci/nlmr/issues/18).

> nlm_gaussianfield

Made all perfect sense, we implemented everything you said
[#19](https://github.com/marcosci/nlmr/issues/19).

> nlm_mosaicfield

We fixed the issue with the loop - [#20](https://github.com/marcosci/nlmr/issues/20).

> nlm_neigh

Again, we changed everything as you recommended :) [#22](https://github.com/marcosci/nlmr/issues/22)

> nlm_polylands

We changed the name of this function, splitted it and got rid of `nlm_randomelement`.
`nlm_randomelement` and `nlm_polylands` both rely on the simulation of point pattern processes
and a spatial interpolation of these patterns.
Since we use the well-established spatstat package to simulate the point patterns, we do not need
`nlm_randomelement` as a direct copy of its NLMpy equivalent and can combine the first option of
`nlm_polylands` and `nlm_randomelement`.

The two new functions are: `nlm_mosaictess` (a unified version of option 1 on
`nlm_polylands` and `nlm_randomelement`) and `nlm_mosaicgibbs`.

In `nlm_mosaicgibbs` we changed the function to control the hardcore process that
creates the spatial point pattern, it should be faster now and takes fewer arguments
to control the output.

All that should be in [#23](https://github.com/marcosci/nlmr/issues/23).

> nlm_randomcluster

We had a look at that function again and completely changed it after your comments.
Again, it was a direct copy as it is done in NLMpy - but the implementation in NLMpy is
actually not the algorithm they cited as reference.

It is now a direct implementation of the random cluster algorithm by Saura et al. (2000),
and you can control for all the parameters they state in their publication.

See our changes in [#33](https://github.com/marcosci/nlmr/issues/33).

> nlm_randomrectangularcluster

Check: [#25](https://github.com/marcosci/nlmr/issues/25)

### responses to @jeffhollist comments

> Vignettes

We had a look on the way `usethis` is doing that (thanks @maelle for the hint),
and for now only the getstarted vignette is part of the package. The rest is only
visible on the homepage.

Craig is currently working on the vignettes and documentation to make it a bit more verbose,
but we already got rid of the raster section.

> Community guidelines

Check:  [#28](https://github.com/marcosci/nlmr/issues/28)

> Package Scope

As stated at the beginning, by splitting the package the scope of them should
be rather streamlined now.

> Future proofing

We changed every function that used `sp` functions to `sf`, there is no direct dependency on `sp` anymore.
However, I see no point at the moment to use `stars`. As far as I know, the future of the `raster` package
is still rather vague - the last thing I have read was that 2D raster should find a unified home in `raster` as a single raster class and more dimensional raster should find their home in `stars`. If I am mistaken here, or you see a way to use `stars` here please let me know.
