### 2018-03-20 908a620
small updates
