+++ R/nlm_neigh.R
@@ -37,7 +37,8 @@
 #' The algorithm uses uniform proportions for each category by default. A vector
 #' with as many proportions as categories and that sums up to 1 can be used for
 #' other distributions.
-#'
+#' @param rescale [\code{logical(1)}]\cr If \code{TRUE} (default), the values
+#'                are rescaled between 0-1.
 #' @return RasterLayer
 #'
 #' @references
@@ -48,7 +49,7 @@
 #'
 #' @examples
 #' # simulate neighborhood model
-#' (neigh_raster <- nlm_neigh(ncol = 20, nrow = 20, p_neigh = 0.1, p_empty = 0.3,
+#' (neigh_raster <- nlm_neigh(ncol = 50, nrow = 50, p_neigh = 0.1, p_empty = 0.3,
 #'                     categories = 5, neighbourhood = 4))
 #'
 #' \dontrun{
@@ -69,7 +70,8 @@ nlm_neigh <-
            p_empty,
            categories = 3,
            neighbourhood = 4,
-           proportions = NA) {
+           proportions = NA,
+           rescale = TRUE) {
 
     # Check function arguments ----
     checkmate::assert_count(ncol, positive = TRUE)
@@ -79,6 +81,8 @@ nlm_neigh <-
     checkmate::assert_count(categories, positive = TRUE)
     checkmate::assert_true(neighbourhood == 4 || neighbourhood == 8)
     checkmate::assert_vector(proportions)
+    checkmate::assert_logical(rescale)
+
     if (!is.na(proportions)) checkmate::assert_true(sum(proportions) == 1)
 
     # Determine cells per categorie
@@ -162,6 +166,11 @@ nlm_neigh <-
                                          ncol(rndneigh_raster) * resolution,
                                          0,
                                          nrow(rndneigh_raster) * resolution)
+
+    # Rescale values to 0-1 ----
+    if (rescale == TRUE) {
+      rndneigh_raster <- util_rescale(rndneigh_raster)
+    }
 
     return(rndneigh_raster)
 }
