+++ man/nlm_fBm.Rd
@@ -1,10 +1,11 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/nlm_fBm.R
-\name{nlm_fBm}
+\name{nlm_fbm}
+\alias{nlm_fbm}
 \alias{nlm_fBm}
-\title{nlm_fBm}
+\title{nlm_fbm}
 \usage{
-nlm_fBm(ncol, nrow, resolution = 1, fract_dim = 1, user_seed = NULL,
+nlm_fbm(ncol, nrow, resolution = 1, fract_dim = 1, user_seed = NULL,
   rescale = TRUE, ...)
 }
 \arguments{
@@ -46,10 +47,10 @@ Neutral landscapes are generated using fractional Brownian motion,
 }
 \examples{
 # simulate fractional brownian motion
-(fBm_raster  <- nlm_fBm(ncol = 20, nrow = 30, fract_dim = 0.8))
+(fbm_raster  <- nlm_fbm(ncol = 20, nrow = 30, fract_dim = 0.8))
 \dontrun{
 # visualize the NLM
-rasterVis::levelplot(fBm_raster, margin = FALSE, par.settings = rasterVis::viridisTheme())
+rasterVis::levelplot(fbm_raster, margin = FALSE, par.settings = rasterVis::viridisTheme())
 }
 }
 \references{
