### 2018-03-20 ed5fe7f
rerun readme and pkgdown ... again
