+++ README.Rmd
@@ -24,15 +24,10 @@ knitr::opts_chunk$set(
 [![](http://cranlogs.r-pkg.org/badges/grand-total/nlmr)](http://cran.rstudio.com/web/packages/NLMR/index.html) 
 [![](https://badges.ropensci.org/188_status.svg)](https://github.com/ropensci/onboarding/issues/188)
 
-# nlmr <img src="vignettes/logo.png" align="right"  height="175" />
 
 `nlmr` is an ``R`` package for simulating **n**eutral **l**andscape **m**odels (NLM). Designed to be a generic framework like [NLMpy](https://pypi.python.org/pypi/nlmpy), it leverages the ability to simulate the most common NLM that are described in the ecological literature. 
 `nlmr` builds on the advantages of the ``raster``-package and returns all simulation as ``RasterLayer``-objects, thus ensuring a direct compability to common GIS tasks and a pretty flexible and simple usage.
 Furthermore, it simulates NLMs within a self-contained, reproducible framework.
-
-<img src="vignettes/nlmr_grid.png"  width="100%" />
-
-<br>
 
 ## Installation
 
