### 2018-03-20 #188 @jhollist
First, I'd be honored to be the mascot!!!  If time allows, would be happy to more than that too.

Second, very happy with the response and direction this is taking.  Two thumbs up on everything and as far as my review is concerned, `nlmr` is good to go.

Again, really nice job.  Makes me happy to see landscape ecology getting better representation in R!
