### 2018-03-21 b6976b2
use macos in travis
