+++ .travis.yml
@@ -2,12 +2,28 @@
 language: R
 sudo: false
 cache: packages
+latex: false
 
 warnings_are_errors: false
 
-r_github_packages:
-  - r-lib/covr
-  - ecohealthalliance/fasterize
+matrix:
+  include:
+    - os: linux
+      env: R_CODECOV=true
+    - os: osx
+      osx_image: xcode9.3beta
+    - os: osx
+      osx_image: xcode9.2
+      before_install:
+        - brew install curl
+        - brew link --force curl
+    - os: osx
+      osx_image: xcode9.2
+    - os: osx
+      osx_image: xcode7.3
+      r_build_args: '--no-build-vignettes'
+      r_check_args: '--ignore-vignettes --no-examples'
+
 
 before_install:
   - sudo add-apt-repository ppa:ubuntugis/ubuntugis-unstable --yes
