+++ appveyor.yml
@@ -14,13 +14,6 @@ cache:
   - C:\RLibrary
 
 # Adapt as necessary starting from here
-environment:
-  USE_RTOOLS: true
-
-build_script:
-  - travis-tool.sh install_github ecohealthalliance/fasterize
-  - travis-tool.sh install_deps
-
 test_script:
   - travis-tool.sh run_tests
 
