+++ appveyor.yml
@@ -14,6 +14,9 @@ cache:
   - C:\RLibrary
 
 # Adapt as necessary starting from here
+environment:
+  USE_RTOOLS: true
+
 test_script:
   - travis-tool.sh run_tests
 
