+++ man/nlm_fBm.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/nlm_fBm.R
+% Please edit documentation in R/nlm_fbm.R
 \name{nlm_fbm}
 \alias{nlm_fbm}
 \alias{nlm_fBm}
