+++ appveyor.yml
@@ -17,6 +17,9 @@ cache:
 environment:
   USE_RTOOLS: true
 
+build_script:
+  - travis-tool.sh install_deps
+
 test_script:
   - travis-tool.sh run_tests
 
