+++ DESCRIPTION
@@ -7,7 +7,7 @@ Authors@R: c(person("Marco", "Sciaini", email = "sciaini.marco@gmail.com", role 
     person("Craig", "Simpkins", email = "simpkinscraig063@gmail.com",
        role =  "aut", comment = "ORCID = '0000-0003-3212-1379'"),
     person("Cédric", "Scherer", email = "cedricphilippscherer@gmail.com", 
-       role =  c("aut")), 
+       role =  c("aut"), comment = "ORCID = '0000-0003-0465-2543'"), 
     person("Laura", "Graham", role = "rev",
        comment = "Laura reviewed the package for rOpenSci, see 
                   https://github.com/ropensci/onboarding/issues/188"),
