+++ .travis.yml
@@ -7,14 +7,8 @@ latex: false
 fortran: false
 warnings_are_errors: false
 
-r:
-  - release
-  - oldrel
-  - devel
-
 matrix:
   include:
-    - os: linux
     - os: osx
       osx_image: xcode9.2
       before_install:
