### 2018-03-21 adfdf68
include linux in travis
