+++ .travis.yml
@@ -9,6 +9,7 @@ warnings_are_errors: false
 
 matrix:
   include:
+    - os: linux
     - os: osx
       osx_image: xcode9.2
       before_install:
