### 2018-03-22 #188 @jhollist
A joke I appreciated!  Also probably to most likely level of involvement I
can promise.  I am begrudgingly well entrenched in mid-career and as such
time to work on things I want to is a rare commodity.

And again, nice job on these packages.  I'll include a mention in a talk I
am working on for US-IALE.

On Thu, Mar 22, 2018 at 6:00 AM, Marco Sciaini <notifications@github.com>
wrote:

> Response to reviews regarding the utility functions
>
> With respect to the comments on slack yesterday, this issue is supposed to
> cover
> now both nlmr and landscapetools.
>
> Again, you can find landscapetools here:
>
> https://github.com/marcosci/landscapetools
> responses to @laurajanegraham <https://github.com/laurajanegraham>
> comments
>
> util_calc_boundaries and util_w2cp
>
> Both of the functions are only internal ones, for the reclassification
> function.
> I can't see any direct use of them - if I am mistaken, we can certainly
> make
> them public.
>
> util_facetplot
>
> You can now specify the number of rows and cols, see #26
> <https://github.com/marcosci/nlmr/issues/26>.
> responses to @jhollist <https://github.com/jhollist> comments
>
> util_import_roboto_condensed
> is this necessary?
>
> I see some value in providing a font with better kerning pairs and
> geometrics numbers
> (compared to the default font ggplot uses). If the dependency on extrafont
> is not worth it
> in your opinion, we can definitely cut that.
>
> ... did I miss anything regarding the utility functions?
> General
>
> As discussed on twitter, I included both of you in the respective
> description of the
> packages - thanks a lot again :)
> And the mascot comment was just a joke - we would definitely benefit from
> your input Jeff :)
>
