+++ README.md
@@ -28,7 +28,6 @@ Example
 Each neutral landscape models is simulated with a single function (all starting with `nlm_`) in `nlmr`, e.g.:
 
 ``` r
-# Simulate 50x50 random cluster raster
 random_cluster <- nlmr::nlm_randomcluster(nrow = 100,
                                       ncol = 100,
                                       p    = 0.5,
@@ -56,7 +55,7 @@ Overview
 | nlm\_curds                    | Simulates a curdled neutral landscape model. Random curdling recursively subdivides the plane into blocks. At each level of the recursion, a fraction of the this block is declared as habitat while the remaining stays matrix. When option `q` is set, it simulates a wheyed curdling model, where previously selected cells that were declared matrix during recursion, can now contain a proportion of habitat cells. | Keitt (2000); Travis and Dytham (2004), Gustafson and Parker (1992) |
 | nlm\_distancegradient         | Simulates a distance gradient neutral landscape model. The function takes the number of columns and rows as input and creates a *RasterLayer* with the same extent. *Origin* is a numeric vector of xmin, xmax, ymin, ymax for a rectangle inside the raster from which the distance is measured.                                                                                                                         | Etherington, Holland, and O’Sullivan (2015)                         |
 | nlm\_edgegradient             | Simulates a linear gradient orientated on a specified or random direction that has a central peak, which runs perpendicular to the gradient direction.                                                                                                                                                                                                                                                                    | Travis and Dytham (2004)                                            |
-| nlm\_fBm                      | Simulates neutral landscapes are generated using fractional Brownian motion, an extension of Brownian motion in which the amount of correlation between steps is controlled by the Hurst coefficient *H*.                                                                                                                                                                                                                 | Schlather et al. (2015)                                             |
+| nlm\_fbm                      | Simulates neutral landscapes are generated using fractional Brownian motion, an extension of Brownian motion in which the amount of correlation between steps is controlled by the Hurst coefficient *H*.                                                                                                                                                                                                                 | Schlather et al. (2015)                                             |
 | nlm\_gaussianfield            | Simulates a spatially correlated random fields (Gaussian random fields) model, where one can control the distance and magnitude of spatial autocorrelatian.                                                                                                                                                                                                                                                               | Schlather et al. (2015)                                             |
 | nlm\_mosaicfield              | Simulates a mosaic random field neutral landscape model.                                                                                                                                                                                                                                                                                                                                                                  | Schlather et al. (2015)                                             |
[…]
