### 2018-03-23 cd44165
General clean up and improvement to the documentation of the nlmr functions
