### 2018-03-23 0ab1a9b
update pkgdown with craigs changes
