### 2018-03-24 #188 @maelle
Thanks a lot for your work @marcosci and thanks @jhollist & @laurajanegraham for very productive reviews! 🎊

A few last comments from me:

* In all three repos, I think it'd make sense to comment a bit on all of them, maybe in a "See also" section.
* Add the pkgdown link to DESCRIPTION, see http://enpiar.com/2017/11/21/getting-down-with-pkgdown/ "if you already have a URL there (say, to your package’s GitHub repository), you can add a second URL to your pkgdown site, separated by a comma"
* In the pkgdown website (of `landscapetools` in particular) you might want to make the grouping of functions in reference more useful. See https://github.com/dirkschumacher/ompr for an example

Now here is the list of things you have to do before I close this issue 😉
- [] Transfer the repos to the rOpenSci organization under "Settings" in your repo.  I have invited you to two teams that should allow you to do so.  You'll be made admin once you do.
- [] Add the review badge to `landscapetools` repo too.
- [] Fix any links in badges for CI and coverage to point to the ropensci URL

Welcome aboard! We'd also love a blog post about your package, either a short-form intro to it (https://ropensci.org/tech-notes/) or long-form post with more narrative about its development. ((https://ropensci.org/blog/). If you are, @stefaniebutland will be in touch about content and timing.
