+++ README.Rmd
@@ -167,6 +167,8 @@ kable(function_tibble) %>%
 * We are very open to contributions - if you are interested check [Contributor Code of Conduct](CONTRIBUTING.md).
     * Please note that this project is released with a [Contributor Code of Conduct](CONDUCT.md). By participating in this project you agree to abide by its terms.
 
+[![ropensci_footer](https://ropensci.org/public_images/github_footer.png)](http://ropensci.org)
+
 ## References
 Etherington, Thomas R., E. Penelope Holland, and David O’Sullivan. 2015. “NLMpy: A Python Software Package for the Creation of Neutral Landscape Models Within a General Numerical Framework.” *Methods in Ecology and Evolution* 6 (2): 164–68.
 
