### 2018-03-24 f0b66b3
followed maelles comments to strictly, first update urls - then transfer repo ...
