+++ DESCRIPTION
@@ -46,8 +46,8 @@ Imports:
     tibble,
     velox,
     sf
-URL: https://marcosci.github.io/nlmr/
-BugReports: https://github.com/marcosci/nlmr/issues/
+URL: https://ropensci.github.io/nlmr/
+BugReports: https://github.com/ropensci/nlmr/issues/
 Suggests:
     testthat,
     covr,
