### 2018-03-25 #188 @marcosci
Yeah 🙌
## @maelle comments
> In all three repos, I think it'd make sense to comment a bit on all of them, maybe in a "See also" section.

You mean crossreferencing nlmr and landscapetools in the readmes of the packages? :)

> Add the pkgdown link to DESCRIPTION, see http://enpiar.com/2017/11/21/getting-down-with-pkgdown/ "if you already have a URL there (say, to your package’s GitHub repository), you can add a second URL to your pkgdown site, separated by a comma"

Done.

> In the pkgdown website (of landscapetools in particular) you might want to make the grouping of functions in reference more useful. See https://github.com/dirkschumacher/ompr for an example

Done.

## @maelle list of things

- [x] Transfer the repos to the rOpenSci organization under "Settings" in your repo. I have invited you to two teams that should allow you to do so. You'll be made admin once you do.
    - Was a bit too ambitious and happy there - transferred nlmr before changing the repo description (changing *marco*sci to *ropen*sci). Could you change that for me ? Sorry. 🙇
- [x] Add the review badge to landscapetools repo too.
- [x] Fix any links in badges for CI and coverage to point to the ropensci URL

And we would be more than happy to write a blog post :smile:
