### 2018-03-25 #188 @maelle
Appveyor badges

* `[![Build status](https://ci.appveyor.com/api/projects/status/djw840fitcvolbxg?svg=true)](https://ci.appveyor.com/project/ropensci/nlmr)`

* `[![Build status](https://ci.appveyor.com/api/projects/status/aehfkxfb5r4vjlm9?svg=true)](https://ci.appveyor.com/project/ropensci/landscapetools)`

Closing this since the blog post discussion can still happen once it's closed. :-)
