### 2018-03-25 #188 @maelle
I forgot two points @marcosci 🙈

* In your .github dir can you also add a CONTRIBUTING.md and a PR template, see https://github.com/ropensci/dotgithubfiles/ for examples

* We're starting to roll out software metadata files to all rOpenSci packages via the Codemeta initiative, see https://github.com/ropensci/codemetar/#codemetar for how to include it in your package, after installing the package - should be easy as running codemetar::write_codemeta() in the root of your package.
