### 2018-03-27 #188 @marcosci
Hi @stefaniebutland and thanks :-) Would the last couple of April be fine to hand in the draft for the blog post?
